moved footer component to components folder, deleted old comments, added new issue for adding the links later when theyre set up
