+++ src/index.tsx
@@ -1,16 +1,12 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import "./index.css";
-import App from "./App";
 import reportWebVitals from "./reportWebVitals";
-import TwoFactorAuthReminders from "./2FA/2FA-Remind";
-import LoggedOutPage from "./pages/loggedOutPage";
 import Login from "./pages/Login";
-import Footer from "./Footer";
+import Footer from "./components/Footer";
 const root = ReactDOM.createRoot(document.getElementById("root")!);
 root.render(
   <React.StrictMode>
-    {/* <TwoFactorAuthReminders /> */}
     <Login />
     <Footer />
   </React.StrictMode>
